web/lsg/requirements.txt
psycopg2
geopy
Pillow
python-social-auth
django>=1.10
xmltodict
djangorestframework
markdown
django-filter
drf-nested-routers
django-pg-fts
jsonfield
django-jsonfield
django-debug-toolbar
admin-extra-urls
eventlet
celery
django-extra-fields
django-activity-stream
git+https://github.com/celery/django-celery.git
thegamesdb

